requirements.txt
asgiref==3.8.1
Django==5.0.8
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
